+++ packages/datatrak-web/src/features/Survey/Components/index.ts
@@ -1,4 +1,5 @@
 export { CancelSurveyConfirmationToken } from './CancelSurveyConfirmationToken';
+export { DraftExistsModal } from './DraftExistsModal';
 export { DesktopSurveyHeader } from './DesktopSurveyHeader';
 export { MobileSurveyHeader } from './MobileSurveyHeader';
 export { MobileSurveyMenu } from './MobileSurveyMenu.tsx';
