+++ packages/datatrak-web/src/features/Survey/index.ts
@@ -1,10 +1,12 @@
 export {
   CancelSurveyConfirmationToken,
   DesktopSurveyHeader,
+  DraftExistsModal,
   MobileSurveyHeader,
   SurveyReviewSection,
   SurveySideMenu,
 } from './Components';
+export { useDraftExistsModal } from './hooks/useDraftExistsModal';
 export * from './Screens';
 export { SurveyContext, getArithmeticDisplayAnswer, useSurveyForm } from './SurveyContext';
 export { SurveyLayout } from './SurveyLayout';
